Move files to create locations
